Support an alternate make command.
On many non-Linux platforms, the vendor-provided make command is not GNU make but either something more POSIXy or something derived from BSD, and GNU make will generally be available under a different name (usually 'gmake').  To support those platforms, allow the name of the make command to be overridden by an environment variable, with the standard 'make' as a fallback.