Always pass -f to tar.
The code that copies the files to be distributed from the source directory into the staging directory assumes that tar, when not provided with a filename, would use stdin / stdout.  This is not the case with POSIXy tars, nor with GNU tar according to its documentation.  Use an explicit -f- instead.