Sort and deduplicate.
Some files were listed twice in this file, which lists the files to include in the distribution tarball.  This caused recent versions of GNU tar to add a spurious hard-link entry for each duplicated file, where previous versions would simply include the whole file twice.  Remove the duplicate entries, and sort the list to help prevent a regression.